backend_data_load/assignment_wk9/requirements.txt
scrapy
neo4j-driver
